chore: push ssa builder into submodule
